+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added delete, patch, post and put commands.
 
 ## [2.0.0] - 2022-10-13
 - Changed required PHP version to >= 8.0.